requirements.txt
# DevOps AI Knowledge Assistant — POC dependencies
# Install: pip install -r requirements.txt

llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openai>=0.2.0
chromadb>=0.4.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
streamlit>=1.30.0
